Extract CommandInvoker
